test(webhook): split the monolithic integration test file by theme
webhook_integration_test.go had grown to ~4.5k lines covering many unrelated
flows. Split the test functions into themed sibling files by build-tag and
package (pure code movement, no test logic changes):

- plan_integration_test.go          — plan and auto-plan flows
- aggregate_check_integration_test.go — aggregate check runs and rerequests
- rollback_integration_test.go       — rollback command flow
- check_cleanup_integration_test.go  — stale cleanup, reconcile, PR close
- no_schema_files_integration_test.go — plan/apply when the PR removes schema files
- webhook_misc_integration_test.go   — metrics and apply dual-write

webhook_integration_test.go now holds only TestMain and the shared setup,
fake-GitHub, and testcontainer helpers used across the package.
