+++ pkg/webhook/webhook_integration_test.go
@@ -1,6 +1,9 @@
 //go:build integration
 
-// Webhook E2E tests exercise the full plan flow end-to-end.
+// Webhook integration tests exercise the full plan flow end to end, against
+// in-process services and MySQL testcontainers (not the e2e/ Docker stack).
+// Shared setup, fake-GitHub wiring, and testcontainer helpers live here; the
+// tests themselves are grouped by theme into sibling *_integration_test.go files.
 //
 // Architecture:
 //
@@ -66,8 +69,6 @@ import (
 	"log"
 	"log/slog"
 	"net/http"
-	"net/http/httptest"
-	"net/url"
 	"os"
 	"os/exec"
 	"strings"
@@ -75,25 +76,14 @@ import (
 	"testing"
 	"time"
 
-	mysql "github.com/go-sql-driver/mysql"
 	gh "github.com/google/go-github/v86/github"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/wait"
 
-	"github.com/block/spirit/pkg/utils"
-
-	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/attribute"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
-	"go.opentelemetry.io/otel/sdk/metric/metricdata"
-
 	"github.com/block/schemabot/pkg/api"
 	ghclient "github.com/block/schemabot/pkg/github"
-	ternv1 "github.com/block/schemabot/pkg/proto/ternv1"
 	"github.com/block/schemabot/pkg/schema"
-	"github.com/block/schemabot/pkg/state"
 	"github.com/block/schemabot/pkg/storage"
 	"github.com/block/schemabot/pkg/storage/mysqlstore"
 	"github.com/block/schemabot/pkg/tern"
@@ -539,1011 +529,6 @@ func registerCompareFiles(t *testing.T, mux *http.ServeMux, baseSHA, headSHA str
 	})
 }
 
-func TestE2EPlanWithChanges(t *testing.T) {
-	dbName := "webhook_plan_changes"
-	svc := setupE2EService(t, dbName)
-	dbConfig := svc.Config().Databases[dbName]
-	dbConfig.AllowedRepos = []string{"octocat/hello-world"}
-	dbConfig.AllowedDirs = []string{"schema"}
-	svc.Config().Databases[dbName] = dbConfig
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "plan generated successfully")
-
-	// Verify plan comment was posted
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "## Schema Change Plan")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, dbName)
-		assert.Contains(t, body, "staging")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for plan comment")
-	}
-
-	// Verify check run was created (per-database + aggregate)
-	select {
-	case cr := <-result.checkRuns:
-		assert.Contains(t, cr.Name, "SchemaBot")
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "action_required", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-
-	// Verify plan was persisted to SchemaBot storage
-	ctx := t.Context()
-	plans, err := svc.Storage().Plans().GetByPR(ctx, "octocat/hello-world", 1)
-	require.NoError(t, err)
-	require.NotEmpty(t, plans, "expected at least one plan record")
-	// Find the plan for this database (shared storage may have data from prior tests)
-	var plan *storage.Plan
-	for _, p := range plans {
-		if p.Database == dbName {
-			plan = p
-			break
-		}
-	}
-	require.NotNil(t, plan, "expected a plan record for database %s", dbName)
-	assert.Equal(t, dbName, plan.Database)
-	assert.Equal(t, "mysql", plan.DatabaseType)
-	assert.Equal(t, "staging", plan.Environment)
-	assert.Equal(t, "octocat/hello-world", plan.Repository)
-	assert.Equal(t, 1, plan.PullRequest)
-	assert.Equal(t, "schema", plan.SchemaPath)
-	assert.NotEmpty(t, plan.PlanIdentifier, "plan should have an identifier")
-	assert.NotNil(t, plan.Namespaces, "plan should have namespace data")
-	assert.NotEmpty(t, plan.Namespaces[dbName].Tables, "plan should have DDL changes")
-
-	// Verify check record was persisted to SchemaBot storage
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check, "expected a check record")
-	assert.Equal(t, "octocat/hello-world", check.Repository)
-	assert.Equal(t, 1, check.PullRequest)
-	assert.Equal(t, "abc123", check.HeadSHA)
-	assert.Equal(t, "staging", check.Environment)
-	assert.Equal(t, "mysql", check.DatabaseType)
-	assert.Equal(t, dbName, check.DatabaseName)
-	assert.True(t, check.HasChanges, "check should indicate changes detected")
-	assert.Equal(t, "completed", check.Status)
-	assert.Equal(t, "action_required", check.Conclusion)
-}
-
-// TestE2EPlanSourcePolicyBlocksUnauthorizedRepo verifies the trusted GitHub
-// discovery path enforces server-side source policy before a plan is stored.
-// The manual command path should also write a failing aggregate check so the
-// Checks UI matches the failure comment.
-func TestE2EPlanSourcePolicyBlocksUnauthorizedRepo(t *testing.T) {
-	dbName := "webhook_source_policy_block"
-	svc := setupE2EService(t, dbName)
-	dbConfig := svc.Config().Databases[dbName]
-	dbConfig.AllowedRepos = []string{"octocat/orders"}
-	dbConfig.AllowedDirs = []string{"schema"}
-	svc.Config().Databases[dbName] = dbConfig
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	h := newE2EHandler(t, svc, client)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "plan failed")
-
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "## ❌ Plan Failed")
-		assert.Contains(t, body, "source policy")
-		assert.Contains(t, body, "repo \"octocat/hello-world\" is not authorized")
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for source policy failure comment")
-	}
-
-	plans, err := svc.Storage().Plans().GetByPR(t.Context(), "octocat/hello-world", 1)
-	require.NoError(t, err)
-	for _, plan := range plans {
-		assert.NotEqual(t, dbName, plan.Database, "source-policy-blocked plan should not be stored")
-	}
-
-	check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	assert.Nil(t, check, "source-policy-blocked plan should not store a per-database check")
-
-	var aggregateCheck checkRunCapture
-	select {
-	case aggregateCheck = <-result.checkRuns:
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for source policy aggregate check run")
-	}
-	assert.Equal(t, aggregateCheckName, aggregateCheck.Name)
-	assert.Equal(t, checkStatusCompleted, aggregateCheck.Status)
-	assert.Equal(t, checkConclusionFailure, aggregateCheck.Conclusion)
-	require.NotNil(t, aggregateCheck.Output)
-	assert.Contains(t, aggregateCheck.Output.Summary, "source policy")
-
-	aggregate, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	require.NotNil(t, aggregate, "source-policy-blocked plan should store a failing aggregate check")
-	assert.Equal(t, "abc123", aggregate.HeadSHA)
-	assert.Equal(t, checkStatusCompleted, aggregate.Status)
-	assert.Equal(t, checkConclusionFailure, aggregate.Conclusion)
-}
-
-func TestE2EPlanNoChanges(t *testing.T) {
-	dbName := "webhook_plan_nochanges"
-	svc := setupE2EService(t, dbName)
-
-	// Create the table in the target DB first so the plan finds no changes
-	ctx := t.Context()
-	appDSN := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName, 1) + "&multiStatements=true"
-	db, err := sql.Open("mysql", appDSN)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "plan generated successfully")
-
-	// Verify plan comment — should say no changes
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "No schema changes detected")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for plan comment")
-	}
-
-	// Verify check run — should be success
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-
-	// Verify plan was persisted to SchemaBot storage
-	plans, err := svc.Storage().Plans().GetByPR(ctx, "octocat/hello-world", 1)
-	require.NoError(t, err)
-	require.NotEmpty(t, plans, "expected at least one plan record")
-	var noChangesPlan *storage.Plan
-	for _, p := range plans {
-		if p.Database == dbName {
-			noChangesPlan = p
-			break
-		}
-	}
-	require.NotNil(t, noChangesPlan, "expected a plan record for database %s", dbName)
-	assert.Equal(t, dbName, noChangesPlan.Database)
-	assert.Equal(t, "staging", noChangesPlan.Environment)
-
-	// Verify check record was persisted — no changes, so conclusion is "success"
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check, "expected a check record")
-	assert.False(t, check.HasChanges, "check should indicate no changes")
-	assert.Equal(t, "completed", check.Status)
-	assert.Equal(t, "success", check.Conclusion)
-}
-
-func TestE2EPlanConfigNotFound(t *testing.T) {
-	dbName := "webhook_plan_noconfig"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// No schemabot.yaml config — empty schema files, no config
-	result := setupFakeGitHubForPlan(t, mux, map[string]string{}, "", dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "schema request error handled")
-
-	// Verify error comment about no config
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "No SchemaBot Configuration Found")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for error comment")
-	}
-}
-
-func TestE2EMultiEnvPlan(t *testing.T) {
-	dbName := "webhook_multi_env"
-	svc := setupE2EServiceMultiEnv(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// "schemabot plan" without -e → multi-env plan
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "multi-env plan started")
-
-	// Multi-env plan runs as a background goroutine — wait for the single combined comment
-	select {
-	case body := <-result.comments:
-		// Should be a combined comment (not separate per env)
-		assert.Contains(t, body, "## Schema Change Plan")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, dbName)
-
-		// Both envs have identical changes (empty target DBs), so should be deduplicated
-		assert.Contains(t, body, "Staging & Production",
-			"identical plans should have combined environment header")
-		assert.NotContains(t, body, "### Staging\n",
-			"should not have separate Staging section when plans are identical")
-
-		// Footer should suggest staging first
-		assert.Contains(t, body, "schemabot apply -e staging")
-		assert.Contains(t, body, "schemabot apply -e production")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for multi-env plan comment")
-	}
-
-	// Should get the aggregate check run
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "action_required", cr.Conclusion)
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for aggregate check run")
-	}
-}
-
-func TestE2EMultiEnvPlanDifferentChanges(t *testing.T) {
-	dbName := "webhook_multi_env_diff"
-	svc := setupE2EServiceMultiEnv(t, dbName)
-
-	// Pre-create the table in staging so staging has no changes, but production still does
-	ctx := t.Context()
-	appDSNStaging := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName+"_staging", 1) + "&multiStatements=true"
-	db, err := sql.Open("mysql", appDSNStaging)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "multi-env plan started")
-
-	// Wait for the combined comment
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "## Schema Change Plan")
-
-		// Plans differ: staging has no changes, production has changes
-		// Should NOT be deduplicated — should show separate sections
-		assert.Contains(t, body, "### Staging")
-		assert.Contains(t, body, "### Production")
-		assert.Contains(t, body, "No schema changes detected")
-		assert.Contains(t, body, "CREATE TABLE")
-
-		// Footer should only suggest production
-		assert.Contains(t, body, "schemabot apply -e production")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for multi-env plan comment")
-	}
-}
-
-func TestE2EAutoPlan(t *testing.T) {
-	dbName := "webhook_autoplan"
-	svc := setupE2EService(t, dbName)
-	svc.Config().Tenant = "alpha"
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// Send a pull_request "opened" webhook instead of an issue_comment
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	// Auto-plan runs in a background goroutine — wait for the plan comment
-	select {
-	case body := <-result.comments:
-		firstLine, _, _ := strings.Cut(body, "\n")
-		assert.Equal(t, "## Schema Change Plan — Staging", firstLine)
-		assert.Contains(t, body, "**Tenant**: `alpha`")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, dbName)
-		assert.Contains(t, body, "schemabot apply -e staging --tenant alpha")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for auto-plan comment")
-	}
-
-	// Verify check run was created
-	select {
-	case cr := <-result.checkRuns:
-		assert.Contains(t, cr.Name, "SchemaBot")
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "action_required", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-}
-
-// A push that only changes non-schema inputs still refreshes SchemaBot checks
-// for the new PR commit without adding another plan comment to the PR timeline.
-func TestE2EAutoPlanSynchronizeApplicationOnlyChangeSkipsComment(t *testing.T) {
-	dbName := "webhook_autoplan_app_only_sync"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-	registerCompareFiles(t, mux, "oldsha", "newsha", []*gh.CommitFile{{
-		Filename: new("app/service.go"),
-		Status:   new("modified"),
-	}})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:    "synchronize",
-		beforeSHA: "oldsha",
-		headSHA:   "newsha",
-		headRef:   "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	select {
-	case cr := <-result.checkRuns:
-		assert.Contains(t, cr.Name, "SchemaBot")
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "action_required", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-
-	select {
-	case body := <-result.comments:
-		t.Fatalf("expected no comment for application-only synchronize, got: %s", body)
-	case <-time.After(3 * time.Second):
-	}
-}
-
-// TestE2EAutoPlanSourcePolicyBlocksWithFailingAggregate verifies auto-plan
-// source-policy failures create a failing aggregate Check Run for branch
-// protection instead of only posting a PR comment.
-func TestE2EAutoPlanSourcePolicyBlocksWithFailingAggregate(t *testing.T) {
-	dbName := "webhook_autoplan_source_policy_block"
-	svc := setupE2EService(t, dbName)
-	dbConfig := svc.Config().Databases[dbName]
-	dbConfig.AllowedRepos = []string{"octocat/orders"}
-	dbConfig.AllowedDirs = []string{"schema"}
-	svc.Config().Databases[dbName] = dbConfig
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "failed to plan")
-		assert.Contains(t, body, "source policy")
-		assert.Contains(t, body, "repo \"octocat/hello-world\" is not authorized")
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for source policy auto-plan comment")
-	}
-
-	var aggregateCheck checkRunCapture
-	select {
-	case aggregateCheck = <-result.checkRuns:
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for source policy aggregate check run")
-	}
-	assert.Equal(t, aggregateCheckName, aggregateCheck.Name)
-	assert.Equal(t, "completed", aggregateCheck.Status)
-	assert.Equal(t, "failure", aggregateCheck.Conclusion)
-	require.NotNil(t, aggregateCheck.Output)
-	assert.Contains(t, aggregateCheck.Output.Summary, "source policy")
-
-	check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	require.NotNil(t, check, "source-policy-blocked auto-plan should store a failing aggregate check")
-	assert.Equal(t, "abc123", check.HeadSHA)
-	assert.Equal(t, "completed", check.Status)
-	assert.Equal(t, "failure", check.Conclusion)
-
-	plans, err := svc.Storage().Plans().GetByPR(t.Context(), "octocat/hello-world", 1)
-	require.NoError(t, err)
-	for _, plan := range plans {
-		assert.NotEqual(t, dbName, plan.Database, "source-policy-blocked auto-plan should not store a plan")
-	}
-}
-
-// TestE2EReopenedPRAutoPlansCurrentHead verifies that reopening a PR follows
-// the same auto-plan path as a new PR and records checks on the current commit.
-func TestE2EReopenedPRAutoPlansCurrentHead(t *testing.T) {
-	dbName := "webhook_reopened_autoplan"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	// Fake GitHub returns schema files and PR metadata for the reopened commit.
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "reopened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "Schema Change Plan")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, dbName)
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for reopened auto-plan comment")
-	}
-
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, "abc123", cr.HeadSHA)
-		assert.Equal(t, checkStatusCompleted, cr.Status)
-		assert.Equal(t, checkConclusionActionRequired, cr.Conclusion)
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for reopened auto-plan check run")
-	}
-
-	// The stored check state must be tied to the reopened commit SHA, not any
-	// stale SHA from before the PR was closed.
-	check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, "abc123", check.HeadSHA)
-	assert.Equal(t, checkConclusionActionRequired, check.Conclusion)
-}
-
-func TestE2EAutoPlanWithLintViolations(t *testing.T) {
-	dbName := "webhook_autoplan_lint"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	// Use a FLOAT column (triggers has_float linter at warning severity, not unsafe).
-	schemaFiles := map[string]string{
-		"bad_table.sql": "CREATE TABLE `bad_table` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `amount` float NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// Send a pull_request "opened" webhook to trigger auto-plan
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	// Wait for the plan comment — should include lint violations from LintSchema
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "Schema Change Plan")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, "Lint Warnings", "plan comment should include lint violations section")
-		assert.Contains(t, body, "bad_table", "lint warning should reference the table name")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for auto-plan comment with lint violations")
-	}
-
-	// Verify check run was created
-	select {
-	case cr := <-result.checkRuns:
-		assert.Contains(t, cr.Name, "SchemaBot")
-		assert.Equal(t, "completed", cr.Status)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-}
-
-func TestE2EAutoPlanNoChangesSkipsComment(t *testing.T) {
-	dbName := "webhook_autoplan_nochange"
-	svc := setupE2EService(t, dbName)
-
-	// Pre-create the table so there are no changes
-	ctx := t.Context()
-	appDSN := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName, 1) + "&multiStatements=true"
-	db, err := sql.Open("mysql", appDSN)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "synchronize",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	// Check run should still be created (for PR status)
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for check run")
-	}
-
-	// No comment should be posted — give it a moment to confirm nothing arrives
-	select {
-	case body := <-result.comments:
-		t.Fatalf("expected no comment for auto-plan with no changes, but got: %s", body)
-	case <-time.After(3 * time.Second):
-		// expected: no comment posted
-	}
-}
-
-func TestE2EAutoPlanNoSchemaFiles(t *testing.T) {
-	dbName := "webhook_autoplan_noschema"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR info
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("abc123"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files — only non-schema files
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, r *http.Request) {
-		files := []*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-			{Filename: new("main.go"), Status: new("modified")},
-		}
-		_ = json.NewEncoder(w).Encode(files)
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "no schema files in PR")
-}
-
-// TestE2EGitHubUnavailableDuringConfigDiscoveryPublishesFailingAggregates
-// verifies that SchemaBot fails closed when it can verify the PR commit but
-// cannot inspect changed files because GitHub returns an availability error.
-func TestE2EGitHubUnavailableDuringConfigDiscoveryPublishesFailingAggregates(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging", "production"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR metadata is available, so SchemaBot knows the current commit SHA and
-	// can safely publish failing aggregate checks against that SHA.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("abc123"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// Changed-file discovery fails after the PR commit is known. This is a
-	// fail-closed condition for every configured environment.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-	})
-
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "config discovery failed")
-
-	seen := map[string]bool{}
-	for i := range 2 {
-		select {
-		case cr := <-checkRuns:
-			seen[cr.Name] = true
-			assert.Equal(t, checkStatusCompleted, cr.Status)
-			assert.Equal(t, checkConclusionFailure, cr.Conclusion)
-			assert.Equal(t, "abc123", cr.HeadSHA)
-		case <-time.After(webhookIntegrationCheckRunDeadline):
-			t.Fatalf("timed out waiting for failing aggregate check run %d/2, seen: %v", i+1, seen)
-		}
-	}
-	assert.True(t, seen["SchemaBot (staging)"])
-	assert.True(t, seen["SchemaBot (production)"])
-
-	// Each aggregate stores a machine-readable GitHub-unavailable blocking
-	// reason so operators can distinguish this from a schema/config error.
-	for _, env := range []string{"staging", "production"} {
-		check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, env, aggregateSentinel, aggregateSentinel)
-		require.NoError(t, err)
-		require.NotNil(t, check)
-		assert.Equal(t, githubConfigDiscoveryUnavailableBlock.blockingReason, check.BlockingReason)
-		assert.Contains(t, check.ErrorMessage, githubConfigDiscoveryUnavailableBlock.message)
-	}
-}
-
-// TestE2EGitHubUnavailableDuringAutoPlanDoesNotPublishCheckRun verifies that
-// SchemaBot does not create or store a check run when it cannot verify the
-// current PR commit SHA at all.
-func TestE2EGitHubUnavailableDuringAutoPlanDoesNotPublishCheckRun(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// The initial PR lookup fails, so SchemaBot does not know which commit SHA
-	// a check run should target.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-	})
-
-	checkRuns := make(chan checkRunCapture, 1)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "config discovery failed")
-
-	// Publishing a check run without a verified head SHA could mark the wrong
-	// commit, so no GitHub or stored aggregate check should be created.
-	select {
-	case cr := <-checkRuns:
-		t.Fatalf("GitHub outage should not publish a check run, got: %+v", cr)
-	case <-time.After(250 * time.Millisecond):
-	}
-
-	aggregate, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	assert.Nil(t, aggregate)
-}
-
 // setupE2EServiceMultiEnv creates a real api.Service with staging and production environments.
 // Each environment gets its own database on the target container.
 func setupE2EServiceMultiEnv(t *testing.T, appDBName string) *api.Service {
@@ -1623,853 +608,6 @@ func setupE2EServiceMultiEnv(t *testing.T, appDBName string) *api.Service {
 
 	return svc
 }
-
-// TestE2EAggregateCheck verifies that a multi-env plan creates a single aggregate
-// "SchemaBot" check run that rolls up per-database checks, and that the aggregate
-// record is persisted in storage with the correct conclusion.
-func TestE2EAggregateCheck(t *testing.T) {
-	dbName := "webhook_aggregate_check"
-	svc := setupE2EServiceMultiEnv(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-	h := newE2EHandler(t, svc, client)
-
-	// Trigger multi-env plan
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Drain the comment
-	select {
-	case <-result.comments:
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for plan comment")
-	}
-
-	// Collect the aggregate check run (only aggregates are created, no per-database)
-	var aggCR checkRunCapture
-	select {
-	case aggCR = <-result.checkRuns:
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for aggregate check run")
-	}
-
-	assert.Equal(t, "SchemaBot", aggCR.Name)
-	assert.Equal(t, "completed", aggCR.Status)
-	assert.Equal(t, "action_required", aggCR.Conclusion)
-
-	// Verify aggregate check record persisted in storage
-	ctx := t.Context()
-	aggCheck, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1,
-		aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	require.NotNil(t, aggCheck, "expected aggregate check record in storage")
-	assert.Equal(t, "action_required", aggCheck.Conclusion)
-	assert.Equal(t, "completed", aggCheck.Status)
-	assert.True(t, aggCheck.HasChanges)
-	assert.Equal(t, "abc123", aggCheck.HeadSHA)
-}
-
-// TestE2EAggregateCheckStaleCleanup verifies that when a new commit removes all schema
-// changes from a PR, the stale per-database checks and aggregate check are re-created
-// on the new HEAD SHA with "success" conclusion. This reproduces the scenario where a
-// user pushes a commit that reverts their schema change.
-func TestE2EAggregateCheckStaleCleanup(t *testing.T) {
-	dbName := "webhook_aggregate_stale"
-	svc := setupE2EServiceMultiEnv(t, dbName)
-	ctx := t.Context()
-
-	// Seed per-database checks and aggregate as if a plan already ran on the first commit.
-	for _, env := range []string{"staging", "production"} {
-		check := &storage.Check{
-			Repository:   "octocat/hello-world",
-			PullRequest:  1,
-			HeadSHA:      "oldsha111",
-			Environment:  env,
-			DatabaseType: "mysql",
-			DatabaseName: dbName,
-			CheckRunID:   100,
-			HasChanges:   true,
-			Status:       checkStatusCompleted,
-			Conclusion:   checkConclusionActionRequired,
-		}
-		require.NoError(t, svc.Storage().Checks().Upsert(ctx, check))
-	}
-	aggCheck := &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  aggregateSentinel,
-		DatabaseType: aggregateSentinel,
-		DatabaseName: aggregateSentinel,
-		CheckRunID:   200,
-		HasChanges:   true,
-		Status:       checkStatusCompleted,
-		Conclusion:   checkConclusionActionRequired,
-	}
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, aggCheck))
-
-	// Set up fake GitHub server that returns NO changed files (simulating revert commit)
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR info with new HEAD SHA
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("newsha222"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// Empty PR files — the revert commit means no schema files changed
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{})
-	})
-
-	// Capture check run creates and updates on the new SHA
-	checkRuns := make(chan checkRunCapture, 10)
-	stalePlanChecksCleaned := func() bool {
-		for _, env := range []string{"staging", "production"} {
-			check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, env, "mysql", dbName)
-			if err != nil || check == nil {
-				return false
-			}
-			if check.HeadSHA != "newsha222" || check.Conclusion != checkConclusionSuccess || check.HasChanges {
-				return false
-			}
-		}
-		return true
-	}
-	var prematurePassingAggregate atomic.Bool
-	var checkRunIDCounter atomic.Int64
-	checkRunIDCounter.Store(300)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		if body.Name == aggregateCheckName && body.Conclusion == checkConclusionSuccess && !stalePlanChecksCleaned() {
-			prematurePassingAggregate.Store(true)
-		}
-		checkRuns <- body
-		id := checkRunIDCounter.Add(1)
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": id})
-	})
-	mux.HandleFunc("PATCH /repos/octocat/hello-world/check-runs/", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		if body.Name == aggregateCheckName && body.Conclusion == checkConclusionSuccess && !stalePlanChecksCleaned() {
-			prematurePassingAggregate.Store(true)
-		}
-		checkRuns <- body
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-
-	// Send synchronize event with new HEAD SHA
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "synchronize",
-		headSHA: "newsha222",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// cleanupStaleChecks marks the plan-only records as success and publishes the
-	// aggregate. The passing-aggregate path should not race ahead while stale
-	// action_required records still exist.
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, checkConclusionSuccess, cr.Conclusion)
-		assert.False(t, prematurePassingAggregate.Load(), "passing aggregate was published before stale per-database records were cleaned")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for aggregate check run")
-	}
-
-	// Poll for per-database storage records to be updated by cleanupStaleChecks.
-	for _, env := range []string{"staging", "production"} {
-		deadline := time.After(5 * time.Second)
-		for {
-			check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, env, "mysql", dbName)
-			if err == nil && check != nil && check.HeadSHA == "newsha222" {
-				assert.Equal(t, checkConclusionSuccess, check.Conclusion)
-				assert.False(t, check.HasChanges)
-				assert.Empty(t, check.BlockingReason)
-				break
-			}
-			select {
-			case <-deadline:
-				t.Fatalf("timed out waiting for %s check to update to new SHA", env)
-			case <-time.After(100 * time.Millisecond):
-			}
-		}
-	}
-
-	// Poll for aggregate storage update.
-	var storedAgg *storage.Check
-	deadline2 := time.After(5 * time.Second)
-	for {
-		storedAgg, _ = svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1,
-			aggregateSentinel, aggregateSentinel, aggregateSentinel)
-		if storedAgg != nil && storedAgg.HeadSHA == "newsha222" {
-			break
-		}
-		select {
-		case <-deadline2:
-			t.Fatal("timed out waiting for aggregate storage update")
-		case <-time.After(100 * time.Millisecond):
-		}
-	}
-	assert.Equal(t, checkConclusionSuccess, storedAgg.Conclusion)
-	assert.False(t, storedAgg.HasChanges)
-	assert.False(t, prematurePassingAggregate.Load(), "passing aggregate was published before stale per-database records were cleaned")
-}
-
-// TestE2ENewHeadPlanPreservesInProgressApplyOwnership verifies the case where
-// an older commit has a running apply, then a newer commit is pushed and
-// auto-planned. The running apply remains authoritative for the stored check
-// state: the new commit's plan result must not take ownership or overwrite it,
-// the aggregate on the new commit stays in_progress (blocking merge), and the
-// apply's own completion still lands its real result.
-func TestE2ENewHeadPlanPreservesInProgressApplyOwnership(t *testing.T) {
-	dbName := "webhook_new_head_replans"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Seed the old commit's running apply. In production this is the apply that
-	// started before the author pushed a newer commit to the PR branch.
-	apply := &storage.Apply{
-		ApplyIdentifier: "apply-old-head",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           state.Apply.Running,
-		Engine:          "spirit",
-	}
-	applyID, err := svc.Storage().Applies().Create(ctx, apply)
-	require.NoError(t, err)
-	apply, err = svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-	require.NotNil(t, apply)
-
-	// Seed old check state owned by that apply. ApplyID makes terminal updates
-	// conditional, so the apply can only complete the check state it owns.
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   100,
-		ApplyID:      applyID,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  aggregateSentinel,
-		DatabaseType: aggregateSentinel,
-		DatabaseName: aggregateSentinel,
-		CheckRunID:   200,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-
-	// Fake GitHub now serves the newer PR commit and schema files. Auto-plan
-	// produces a plan result for the new commit, but the running apply keeps
-	// ownership of the stored check state.
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "synchronize",
-		headSHA: "abc123",
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// The aggregate is re-created on the new commit and stays in_progress
-	// because the apply-owned per-database row still blocks it.
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, "abc123", cr.HeadSHA)
-		assert.Equal(t, checkStatusInProgress, cr.Status)
-		assert.Empty(t, cr.Conclusion)
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for new-head aggregate check run")
-	}
-
-	// The running apply keeps ownership: the new commit's plan result must not
-	// overwrite the in-progress check state or clear the apply ID.
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, "oldsha111", check.HeadSHA)
-	assert.Equal(t, checkStatusInProgress, check.Status)
-	assert.Empty(t, check.Conclusion)
-	assert.Equal(t, applyID, check.ApplyID)
-
-	// The apply still owns the check state, so its completion lands the real
-	// terminal result instead of missing ownership.
-	apply.State = state.Apply.Completed
-	updated, err := h.updateCheckRecordForApplyResult(ctx, "octocat/hello-world", 1, apply)
-	require.NoError(t, err)
-	assert.True(t, updated, "apply completion should update the check state it owns")
-
-	check, err = svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, checkStatusCompleted, check.Status)
-	assert.Equal(t, checkConclusionSuccess, check.Conclusion)
-	assert.Equal(t, applyID, check.ApplyID)
-}
-
-func TestE2EAggregateCheckStaleCleanupBlocksStartedApply(t *testing.T) {
-	dbName := "webhook_aggregate_stale_apply"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	apply := &storage.Apply{
-		ApplyIdentifier: "apply-reverted-commit",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           state.Apply.Running,
-		Engine:          "spirit",
-	}
-	applyID, err := svc.Storage().Applies().Create(ctx, apply)
-	require.NoError(t, err)
-	apply, err = svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-	require.NotNil(t, apply)
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   100,
-		ApplyID:      applyID,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  aggregateSentinel,
-		DatabaseType: aggregateSentinel,
-		DatabaseName: aggregateSentinel,
-		CheckRunID:   200,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("newsha222"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{})
-	})
-
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 300})
-	})
-	mux.HandleFunc("PATCH /repos/octocat/hello-world/check-runs/", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 300})
-	})
-
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "synchronize",
-		headSHA: "newsha222",
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, "newsha222", cr.HeadSHA)
-		assert.Equal(t, checkStatusInProgress, cr.Status)
-		assert.Empty(t, cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for in-progress aggregate check run")
-	}
-
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-	h.postPassingAggregates(ctx, installClient, "octocat/hello-world", 1, "newsha222")
-	select {
-	case cr := <-checkRuns:
-		require.NotEqual(t, checkConclusionSuccess, cr.Conclusion, "passing aggregate must not be published while a started apply blocks the PR")
-	case <-time.After(250 * time.Millisecond):
-	}
-
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, "newsha222", check.HeadSHA)
-	assert.Equal(t, checkStatusInProgress, check.Status)
-	assert.Empty(t, check.Conclusion)
-	assert.True(t, check.HasChanges)
-	assert.Equal(t, applyID, check.ApplyID)
-	assert.Equal(t, schemaRemovedAfterApplyBlock.blockingReason, check.BlockingReason)
-	assert.Equal(t, schemaRemovedAfterApplyBlock.message, check.ErrorMessage)
-
-	apply.State = state.Apply.Completed
-	updated, err := h.updateCheckRecordForApplyResult(ctx, "octocat/hello-world", 1, apply)
-	require.NoError(t, err)
-	assert.True(t, updated, "old apply completion should finish the owning row without marking it successful")
-
-	check, err = svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, checkStatusCompleted, check.Status)
-	assert.Equal(t, checkConclusionActionRequired, check.Conclusion)
-	assert.Equal(t, schemaRemovedAfterApplyBlock.blockingReason, check.BlockingReason)
-	assert.Equal(t, schemaRemovedAfterApplyBlock.message, check.ErrorMessage)
-
-	h.updateAggregateCheck(ctx, installClient, "octocat/hello-world", 1, "newsha222")
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, aggregateCheckName, cr.Name)
-		assert.Equal(t, checkStatusCompleted, cr.Status)
-		assert.Equal(t, checkConclusionActionRequired, cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for terminal blocking aggregate check run")
-	}
-}
-
-// TestE2EPlanWithNoCurrentSchemaFilesExplainsInProgressApply verifies that a
-// PR whose current diff no longer contains managed schema files does not fall
-// back to whole-repo config discovery when an apply from that PR is still
-// running. The command should tell the user how to reconcile the in-flight apply.
-func TestE2EPlanWithNoCurrentSchemaFilesExplainsInProgressApply(t *testing.T) {
-	body, treeCalls := runCommandWithNoCurrentSchemaFilesAndApplyOwnedCheck(t, "schemabot plan -e staging", state.Apply.Running, checkStatusInProgress, "")
-
-	assert.Zero(t, treeCalls, "empty-diff plan should not use whole-repo git tree discovery")
-	assert.Contains(t, body, "SchemaBot is still applying a schema change from this PR")
-	assert.Contains(t, body, "The live database operation was already started")
-	assert.Contains(t, body, "schemabot stop apply-empty-diff -e staging")
-	assert.Contains(t, body, "schemabot rollback apply-empty-diff -e staging")
-	assert.Contains(t, body, "schemabot plan -e staging -d webhook_empty_diff_apply")
-	assert.Contains(t, body, "push a no-op `schemabot.yaml` edit to trigger a fresh plan")
-	assert.NotContains(t, body, "ask an operator")
-	assert.NotContains(t, body, "truncated repository tree")
-	assert.NotContains(t, body, "schemabot status")
-}
-
-// TestE2EPlanWithNoCurrentSchemaFilesExplainsCompletedApply verifies that a PR
-// whose current diff no longer contains managed schema files produces a clear
-// reconciliation-required comment after the apply has reached a terminal state.
-func TestE2EPlanWithNoCurrentSchemaFilesExplainsCompletedApply(t *testing.T) {
-	body, treeCalls := runCommandWithNoCurrentSchemaFilesAndApplyOwnedCheck(t, "schemabot plan -e staging", state.Apply.Completed, checkStatusCompleted, checkConclusionActionRequired)
-
-	assert.Zero(t, treeCalls, "empty-diff plan should not use whole-repo git tree discovery")
-	assert.Contains(t, body, "SchemaBot already applied a schema change from this PR")
-	assert.Contains(t, body, "The live database was already updated")
-	assert.Contains(t, body, "Keep the live schema change")
-	assert.Contains(t, body, "Undo the live schema change")
-	assert.Contains(t, body, "schemabot rollback apply-empty-diff -e staging")
-	assert.Contains(t, body, "schemabot plan -e staging -d webhook_empty_diff_apply")
-	assert.Contains(t, body, "push a no-op `schemabot.yaml` edit to trigger a fresh plan")
-	assert.NotContains(t, body, "ask an operator")
-	assert.NotContains(t, body, "truncated repository tree")
-	assert.NotContains(t, body, "Git reverting")
-}
-
-// TestE2EAutoPlanWithOnlySchemaBotConfigChangeClearsRollbackCheck verifies the
-// self-serve reconciliation path for a PR whose live database already matches
-// the current schema files. A no-op schemabot.yaml edit should make config
-// discovery deterministic, auto-plan no changes, and clear the
-// rollback-created blocking check state without operator intervention.
-func TestE2EAutoPlanWithOnlySchemaBotConfigChangeClearsRollbackCheck(t *testing.T) {
-	dbName := "webhook_noop_config_reconcile"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	schemaSQL := "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	cfg, err := mysql.ParseDSN(e2eTargetDSN)
-	require.NoError(t, err)
-	cfg.DBName = dbName
-	db, err := sql.Open("mysql", cfg.FormatDSN())
-	require.NoError(t, err)
-	defer utils.CloseAndLog(db)
-	require.NoError(t, db.PingContext(ctx))
-	_, err = db.ExecContext(ctx, strings.TrimSuffix(schemaSQL, ";"))
-	require.NoError(t, err)
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:     "octocat/hello-world",
-		PullRequest:    1,
-		HeadSHA:        "abc123",
-		Environment:    "staging",
-		DatabaseType:   "mysql",
-		DatabaseName:   dbName,
-		CheckRunID:     100,
-		HasChanges:     true,
-		Status:         checkStatusCompleted,
-		Conclusion:     checkConclusionActionRequired,
-		BlockingReason: rollbackCompletedBlock.blockingReason,
-		ErrorMessage:   rollbackCompletedBlock.message,
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	setupFakeGitHubForPlanWithPRFiles(t, mux, map[string]string{"users.sql": schemaSQL}, schemabotConfig, dbName, []*gh.CommitFile{
-		{
-			Filename: new("schema/schemabot.yaml"),
-			Status:   new("modified"),
-		},
-	})
-
-	h := newE2EHandler(t, svc, client)
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{action: "synchronize"}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	var check *storage.Check
-	require.Eventually(t, func() bool {
-		check, err = svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-		return err == nil && check != nil && check.Status == checkStatusCompleted && check.Conclusion == checkConclusionSuccess
-	}, webhookIntegrationPollDeadline, 500*time.Millisecond, "no-op config auto-plan should clear rollback-created blocking check")
-	assert.Equal(t, "abc123", check.HeadSHA)
-	assert.False(t, check.HasChanges)
-	assert.Zero(t, check.ApplyID)
-	assert.Empty(t, check.BlockingReason)
-	assert.Empty(t, check.ErrorMessage)
-}
-
-// TestE2EApplyWithNoCurrentSchemaFilesExplainsInProgressApply verifies that
-// apply uses the same reconciliation guard as plan instead of falling back to
-// whole-repo config discovery after the current PR diff no longer contains
-// managed schema files.
-func TestE2EApplyWithNoCurrentSchemaFilesExplainsInProgressApply(t *testing.T) {
-	body, treeCalls := runCommandWithNoCurrentSchemaFilesAndApplyOwnedCheck(t, "schemabot apply -e staging", state.Apply.Running, checkStatusInProgress, "")
-
-	assert.Zero(t, treeCalls, "empty-diff apply should not use whole-repo git tree discovery")
-	assert.Contains(t, body, "SchemaBot is still applying a schema change from this PR")
-	assert.Contains(t, body, "schemabot rollback apply-empty-diff -e staging")
-	assert.NotContains(t, body, "truncated repository tree")
-}
-
-// TestE2EApplyConfirmWithNoCurrentSchemaFilesExplainsInProgressApply verifies
-// that apply-confirm also fails closed before discovery when a PR no longer has
-// managed schema files but still owns an in-flight apply.
-func TestE2EApplyConfirmWithNoCurrentSchemaFilesExplainsInProgressApply(t *testing.T) {
-	body, treeCalls := runCommandWithNoCurrentSchemaFilesAndApplyOwnedCheck(t, "schemabot apply-confirm -e staging", state.Apply.Running, checkStatusInProgress, "")
-
-	assert.Zero(t, treeCalls, "empty-diff apply-confirm should not use whole-repo git tree discovery")
-	assert.Contains(t, body, "SchemaBot is still applying a schema change from this PR")
-	assert.Contains(t, body, "schemabot stop apply-empty-diff -e staging")
-	assert.Contains(t, body, "schemabot rollback apply-empty-diff -e staging")
-	assert.NotContains(t, body, "truncated repository tree")
-}
-
-func runCommandWithNoCurrentSchemaFilesAndApplyOwnedCheck(t *testing.T, comment, applyState, checkStatus, checkConclusion string) (string, int64) {
-	t.Helper()
-
-	dbName := "webhook_empty_diff_apply"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	apply := &storage.Apply{
-		ApplyIdentifier: "apply-empty-diff",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           applyState,
-		Engine:          "spirit",
-	}
-	applyID, err := svc.Storage().Applies().Create(ctx, apply)
-	require.NoError(t, err)
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   100,
-		ApplyID:      applyID,
-		HasChanges:   true,
-		Status:       checkStatus,
-		Conclusion:   checkConclusion,
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("newsha222"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{})
-	})
-
-	var treeCalls atomic.Int64
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/", func(w http.ResponseWriter, _ *http.Request) {
-		treeCalls.Add(1)
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA:       new("newsha222"),
-			Entries:   []*gh.TreeEntry{},
-			Truncated: new(true),
-		})
-	})
-
-	comments := make(chan string, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/1/comments", func(w http.ResponseWriter, r *http.Request) {
-		var body struct {
-			Body string `json:"body"`
-		}
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		comments <- body.Body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 99})
-	})
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/comments/42/reactions", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: comment,
-		isPR:    true,
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case body := <-comments:
-		return body, treeCalls.Load()
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for reconciliation comment")
-		return "", treeCalls.Load()
-	}
-}
-
-// TestE2EPlanUsesServerSideTarget verifies that the webhook plan handler routes
-// using the database target policy from server config.
-func TestE2EPlanUsesServerSideTarget(t *testing.T) {
-	dbName := "webhook_server_target"
-	ctx := t.Context()
-
-	// Create the app database on the target
-	targetDB, err := sql.Open("mysql", e2eTargetDSN+"&multiStatements=true")
-	require.NoError(t, err)
-	_, err = targetDB.ExecContext(ctx, "CREATE DATABASE IF NOT EXISTS `"+dbName+"`")
-	require.NoError(t, err)
-	_ = targetDB.Close()
-
-	t.Cleanup(func() {
-		db, err := sql.Open("mysql", e2eTargetDSN+"&multiStatements=true")
-		if err == nil {
-			_, _ = db.ExecContext(t.Context(), "DROP DATABASE IF EXISTS `"+dbName+"`")
-			_ = db.Close()
-		}
-	})
-
-	appDSN := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName, 1)
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-
-	schemabotDB, err := sql.Open("mysql", e2eSchemabotDSN)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = schemabotDB.Close() })
-
-	st := mysqlstore.New(schemabotDB)
-
-	// Clean up stale data
-	_, _ = schemabotDB.ExecContext(ctx, "DELETE FROM checks WHERE database_name = ?", dbName)
-	_, _ = schemabotDB.ExecContext(ctx, "DELETE FROM checks WHERE repository = 'octocat/hello-world' AND pull_request = 1")
-	_, _ = schemabotDB.ExecContext(ctx, "DELETE FROM plans WHERE database_name = ?", dbName)
-
-	localClient, err := tern.NewLocalClient(tern.LocalConfig{
-		Database:  dbName,
-		Type:      "mysql",
-		TargetDSN: appDSN,
-	}, st, logger)
-	require.NoError(t, err)
-	t.Cleanup(func() { _ = localClient.Close() })
-
-	// The tern client is registered under "team-a/staging", so plan must use
-	// the deployment stored in databases.<db>.environments.staging.
-	serverConfig := &api.ServerConfig{
-		Databases: map[string]api.DatabaseConfig{
-			dbName: {
-				Type: "mysql",
-				Environments: map[string]api.EnvironmentConfig{
-					"staging": {Target: "team-a-target", Deployment: "team-a"},
-				},
-			},
-		},
-		TernDeployments: api.TernConfig{
-			"team-a": api.TernEndpoints{
-				"staging": "localhost:9999", // address not dialed; pre-injected client is used instead
-			},
-		},
-		Repos: map[string]api.RepoConfig{
-			"octocat/hello-world": {},
-		},
-	}
-
-	svc := api.New(st, serverConfig, map[string]tern.Client{
-		"team-a/staging": localClient,
-	}, logger)
-	t.Cleanup(func() { _ = svc.Close() })
-
-	// Set up fake GitHub API
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "plan generated successfully")
-
-	// Verify plan comment was posted with the expected DDL
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "## Schema Change Plan")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, dbName)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for plan comment")
-	}
-}
-
-// --- Container helpers (matches e2e/setup_test.go patterns) ---
 
 func startE2EMySQLContainer(ctx context.Context, baseName, dbName string, schemaFS *embed.FS) (testcontainers.Container, error) {
 	req := testcontainers.ContainerRequest{
@@ -2562,1122 +700,6 @@ func applyEmbeddedSchema(db *sql.DB, schemaFS embed.FS) error {
 	return nil
 }
 
-// TestE2EApplyCreateDualWritesApplyOperationRow verifies the service-level
-// apply create path writes exactly one apply_operations row in the same
-// transaction as the applies row, mirroring the apply's (deployment, target)
-// routing. The operator claim loop is not yet wired to these rows, so the
-// row stays in pending — that's the contract until a subsequent PR lifts
-// the deployments-map gate and introduces the per-row claim loop.
-func TestE2EApplyCreateDualWritesApplyOperationRow(t *testing.T) {
-	dbName := "webhook_apply_dual_write"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Seed the target so the plan produces a real DDL change.
-	appDSN := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName, 1) + "&multiStatements=true"
-	db, err := sql.Open("mysql", appDSN)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	schemaWithIndex := "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`),\n  KEY `idx_name` (`name`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	planResp, err := svc.ExecutePlan(ctx, api.PlanRequest{
-		Database:    dbName,
-		Environment: "staging",
-		Type:        "mysql",
-		SchemaFiles: map[string]*ternv1.SchemaFiles{
-			dbName: {Files: map[string]string{"users.sql": schemaWithIndex}},
-		},
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, planResp.Changes, "expected DDL changes")
-
-	applyResp, applyID, err := svc.ExecuteApply(ctx, api.ApplyRequest{
-		PlanID:      planResp.PlanID,
-		Environment: "staging",
-		Options:     map[string]string{"allow_unsafe": "true"},
-	})
-	require.NoError(t, err)
-	require.True(t, applyResp.Accepted)
-	require.Greater(t, applyID, int64(0))
-
-	apply, err := svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-	require.NotNil(t, apply)
-	plan, err := svc.Storage().Plans().Get(ctx, planResp.PlanID)
-	require.NoError(t, err)
-	require.NotNil(t, plan)
-
-	ops, err := svc.Storage().ApplyOperations().ListByApply(ctx, applyID)
-	require.NoError(t, err)
-	require.Len(t, ops, 1, "apply create must dual-write exactly one apply_operations row")
-	op := ops[0]
-	assert.Equal(t, applyID, op.ApplyID)
-	assert.Equal(t, apply.Deployment, op.Deployment, "operation deployment must match apply deployment")
-	assert.Equal(t, plan.Target, op.Target, "operation target must mirror the plan-time target")
-	assert.Equal(t, state.ApplyOperation.Pending, op.State, "operation row stays pending — no consumer is wired yet")
-}
-
-// TestE2ERollbackPlanViaWebhook tests the full rollback flow:
-// 1. Plan + apply a schema change via the service (simulating a prior apply)
-// 2. Run "schemabot rollback <apply-id> -e staging" via webhook
-// 3. Verify the rollback plan comment is posted with reverse DDL
-func TestE2ERollbackPlanViaWebhook(t *testing.T) {
-	dbName := "webhook_rollback"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Step 1: Create an initial table in the target DB (the "before" state)
-	appDSN := strings.Replace(e2eTargetDSN, "/target_test", "/"+dbName, 1) + "&multiStatements=true"
-	db, err := sql.Open("mysql", appDSN)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	// Step 2: Plan + apply adding an index (this stores original files for rollback)
-	schemaWithIndex := "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`),\n  KEY `idx_name` (`name`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	prNumber := int32(1)
-	planReq := api.PlanRequest{
-		Database:    dbName,
-		Environment: "staging",
-		Type:        "mysql",
-		Repository:  "octocat/hello-world",
-		PullRequest: &prNumber,
-		SchemaFiles: map[string]*ternv1.SchemaFiles{
-			dbName: {Files: map[string]string{"users.sql": schemaWithIndex}},
-		},
-	}
-	planResp, err := svc.ExecutePlan(ctx, planReq)
-	require.NoError(t, err)
-	require.NotEmpty(t, planResp.Changes, "expected DDL changes")
-
-	applyReq := api.ApplyRequest{
-		PlanID:      planResp.PlanID,
-		Environment: "staging",
-		Options:     map[string]string{"allow_unsafe": "true"},
-	}
-	applyResp, applyID, err := svc.ExecuteApply(ctx, applyReq)
-	require.NoError(t, err)
-	require.True(t, applyResp.Accepted)
-	require.Greater(t, applyID, int64(0))
-
-	// Wait for apply to complete
-	require.Eventually(t, func() bool {
-		apply, err := svc.Storage().Applies().Get(ctx, applyID)
-		if err != nil || apply == nil {
-			return false
-		}
-		return state.IsState(apply.State, state.Apply.Completed)
-	}, 30*time.Second, 500*time.Millisecond, "apply should complete")
-
-	// Step 3: Set up fake GitHub and webhook handler
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	// Schema files still have the index (current desired state)
-	result := setupFakeGitHubForPlan(t, mux, map[string]string{
-		"users.sql": schemaWithIndex,
-	}, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// Get the apply identifier for the rollback command
-	storedApply, err := svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-	require.NotNil(t, storedApply)
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "abc123",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   42,
-		ApplyID:      applyID,
-		HasChanges:   false,
-		Status:       checkStatusCompleted,
-		Conclusion:   checkConclusionSuccess,
-	}))
-
-	// Step 4: Send rollback command with the apply ID
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: fmt.Sprintf("schemabot rollback %s -e staging", storedApply.ApplyIdentifier),
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "rollback started")
-
-	// Step 5: Verify rollback plan comment was posted
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "## Schema Rollback Plan")
-		assert.Contains(t, body, "DROP INDEX", "rollback should drop the index we added")
-		assert.Contains(t, body, "schemabot rollback-confirm -e staging")
-		assert.Contains(t, body, "schemabot unlock")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for rollback plan comment")
-	}
-
-	// Step 6: Verify lock was acquired
-	lock, err := svc.Storage().Locks().Get(ctx, dbName, "mysql")
-	require.NoError(t, err)
-	require.NotNil(t, lock, "lock should be held after rollback command")
-	assert.Equal(t, "octocat/hello-world#1", lock.Owner)
-	assert.True(t, strings.HasPrefix(lock.PendingPlanID, rollbackPendingPlanPrefix), "rollback lock should pin a tagged rollback plan")
-
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, checkStatusCompleted, check.Status)
-	assert.Equal(t, checkConclusionSuccess, check.Conclusion)
-	assert.Equal(t, applyID, check.ApplyID)
-
-	select {
-	case cr := <-result.checkRuns:
-		t.Fatalf("rollback planning should not update check runs before confirmation, got: %+v", cr)
-	case <-time.After(500 * time.Millisecond):
-	}
-}
-
-// TestE2ERollbackApplyNotFound tests rollback with a nonexistent apply ID.
-func TestE2ERollbackApplyNotFound(t *testing.T) {
-	dbName := "webhook_rollback_none"
-	svc := setupE2EService(t, dbName)
-
-	h, comments, _ := newTestHandler(t)
-	h.service = svc
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot rollback apply_deadbeef0000 -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case body := <-comments:
-		assert.Contains(t, body, "Apply Not Found")
-		assert.Contains(t, body, "apply_deadbeef0000")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for comment")
-	}
-}
-
-// TestE2ERollbackConfirmNoLock tests rollback-confirm when no lock is held.
-func TestE2ERollbackConfirmNoLock(t *testing.T) {
-	dbName := "webhook_rbconfirm_nolock"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	result := setupFakeGitHubForPlan(t, mux, map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	_, err := svc.Storage().Applies().Create(t.Context(), &storage.Apply{
-		ApplyIdentifier: "apply_aabbccdd0012",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           "completed",
-		Engine:          "spirit",
-	})
-	require.NoError(t, err)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot rollback-confirm -e staging",
-		isPR:    true,
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Should post a "no lock found" comment
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "No Lock Found")
-		assert.Contains(t, body, "schemabot rollback <apply-id> -e staging")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for no-lock comment")
-	}
-}
-
-// TestE2ERollbackConfirmExecutesAndPostsComments verifies the full rollback-confirm
-// flow: rollback plan → rollback-confirm → apply executes → summary comment posted
-// on the correct PR. This catches regressions where watchApplyProgress loses the
-// repo/PR/installationID context and fails to post comments.
-func TestE2ERollbackConfirmExecutesAndPostsComments(t *testing.T) {
-	dbName := "webhook_rbconfirm_exec"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Step 1: Create initial table
-	cfg, err := mysql.ParseDSN(e2eTargetDSN)
-	require.NoError(t, err)
-	cfg.DBName = dbName
-	cfg.MultiStatements = true
-	appDSN := cfg.FormatDSN()
-	db, err := sql.Open("mysql", appDSN)
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	// Step 2: Plan + apply adding an index (captures original files for rollback)
-	schemaWithIndex := "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`),\n  KEY `idx_name` (`name`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	prNumber := int32(1)
-	planResp, err := svc.ExecutePlan(ctx, api.PlanRequest{
-		Database:    dbName,
-		Environment: "staging",
-		Type:        "mysql",
-		Repository:  "octocat/hello-world",
-		PullRequest: &prNumber,
-		SchemaFiles: map[string]*ternv1.SchemaFiles{
-			dbName: {Files: map[string]string{"users.sql": schemaWithIndex}},
-		},
-	})
-	require.NoError(t, err)
-
-	applyResp, applyID, err := svc.ExecuteApply(ctx, api.ApplyRequest{
-		PlanID:      planResp.PlanID,
-		Environment: "staging",
-		Options:     map[string]string{"allow_unsafe": "true"},
-	})
-	require.NoError(t, err)
-	require.True(t, applyResp.Accepted)
-
-	require.Eventually(t, func() bool {
-		a, err := svc.Storage().Applies().Get(ctx, applyID)
-		return err == nil && a != nil && a.State == "completed"
-	}, 30*time.Second, 500*time.Millisecond, "initial apply should complete")
-
-	// Step 3: Run rollback to generate plan and acquire lock
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	result := setupFakeGitHubForPlan(t, mux, map[string]string{
-		"users.sql": schemaWithIndex,
-	}, schemabotConfig, dbName)
-
-	h := newE2EHandler(t, svc, client)
-
-	storedApply, err := svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: fmt.Sprintf("schemabot rollback %s -e staging", storedApply.ApplyIdentifier),
-		isPR:    true,
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Drain the rollback plan comment
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "Rollback Plan")
-	case <-time.After(30 * time.Second):
-		t.Fatal("timed out waiting for rollback plan comment")
-	}
-
-	// Step 4: Run rollback-confirm — this triggers the apply + watchApplyProgress
-	req = buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot rollback-confirm -e staging",
-		isPR:    true,
-	}, nil)
-	rr = httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Step 5: Verify that the summary comment arrives on the PR.
-	// This is the critical assertion — if repo/PR/installationID are wrong,
-	// the comment goes to the wrong URL and never reaches the channel.
-	gotSummary := false
-	deadline := time.After(webhookIntegrationPollDeadline)
-	for !gotSummary {
-		select {
-		case body := <-result.comments:
-			if strings.Contains(body, "Schema Change") && (strings.Contains(body, "Applied") || strings.Contains(body, "Complete") || strings.Contains(body, "Failed")) {
-				gotSummary = true
-				assert.Contains(t, body, "DROP INDEX", "rollback should drop the index")
-			}
-		case <-deadline:
-			t.Fatal("timed out waiting for rollback summary comment — " +
-				"watchApplyProgress may have lost repo/PR/installationID context")
-		}
-	}
-}
-
-// TestE2ERollbackConfirmUpdatesCheckToActionRequired verifies that after a
-// rollback-confirm completes, the check run transitions to action_required
-// (not success) since the PR's schema changes have been undone.
-func TestE2ERollbackConfirmUpdatesCheckToActionRequired(t *testing.T) {
-	dbName := "webhook_rb_check"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Step 1: Create initial table
-	cfg, err := mysql.ParseDSN(e2eTargetDSN)
-	require.NoError(t, err)
-	cfg.DBName = dbName
-	cfg.MultiStatements = true
-	db, err := sql.Open("mysql", cfg.FormatDSN())
-	require.NoError(t, err)
-	_, err = db.ExecContext(ctx, "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci")
-	require.NoError(t, err)
-	_ = db.Close()
-
-	// Step 2: Plan + apply adding an index
-	schemaWithIndex := "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`),\n  KEY `idx_name` (`name`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	prNumber := int32(1)
-	planResp, err := svc.ExecutePlan(ctx, api.PlanRequest{
-		Database:    dbName,
-		Environment: "staging",
-		Type:        "mysql",
-		Repository:  "octocat/hello-world",
-		PullRequest: &prNumber,
-		SchemaFiles: map[string]*ternv1.SchemaFiles{
-			dbName: {Files: map[string]string{"users.sql": schemaWithIndex}},
-		},
-	})
-	require.NoError(t, err)
-
-	applyResp, applyID, err := svc.ExecuteApply(ctx, api.ApplyRequest{
-		PlanID:      planResp.PlanID,
-		Environment: "staging",
-		Options:     map[string]string{"allow_unsafe": "true"},
-	})
-	require.NoError(t, err)
-	require.True(t, applyResp.Accepted)
-
-	require.Eventually(t, func() bool {
-		a, err := svc.Storage().Applies().Get(ctx, applyID)
-		return err == nil && a != nil && a.State == "completed"
-	}, 30*time.Second, 500*time.Millisecond, "initial apply should complete")
-
-	// Step 3: Seed a check record (simulates what plan/apply creates)
-	err = svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "abc123",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   42,
-		ApplyID:      applyID,
-		HasChanges:   false,
-		Status:       checkStatusCompleted,
-		Conclusion:   checkConclusionSuccess,
-	})
-	require.NoError(t, err)
-
-	// Step 4: Set up fake GitHub and run rollback + rollback-confirm
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	result := setupFakeGitHubForPlan(t, mux, map[string]string{
-		"users.sql": schemaWithIndex,
-	}, schemabotConfig, dbName)
-
-	h := newE2EHandler(t, svc, client)
-
-	storedApply, err := svc.Storage().Applies().Get(ctx, applyID)
-	require.NoError(t, err)
-
-	// Run rollback
-	req := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: fmt.Sprintf("schemabot rollback %s -e staging", storedApply.ApplyIdentifier),
-		isPR:    true,
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case <-result.comments:
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for rollback plan comment")
-	}
-
-	// Run rollback-confirm
-	req = buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot rollback-confirm -e staging",
-		isPR:    true,
-	}, nil)
-	rr = httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Wait for the rollback apply to complete and check to be updated
-	require.Eventually(t, func() bool {
-		check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-		if err != nil {
-			return false
-		}
-		return isRollbackActionRequiredWithoutApplyOwnership(check)
-	}, webhookIntegrationPollDeadline, 500*time.Millisecond,
-		"check should transition to action_required without active apply ownership after rollback")
-
-	deadline := time.After(webhookIntegrationPollDeadline)
-	for {
-		select {
-		case cr := <-result.checkRuns:
-			if cr.Name == aggregateCheckName &&
-				cr.Status == checkStatusCompleted &&
-				cr.Conclusion == checkConclusionActionRequired {
-				return
-			}
-		case <-deadline:
-			t.Fatal("timed out waiting for rollback aggregate to become action_required")
-		}
-	}
-}
-
-func isRollbackActionRequiredWithoutApplyOwnership(check *storage.Check) bool {
-	if check == nil {
-		return false
-	}
-	return check.Status == checkStatusCompleted &&
-		check.Conclusion == checkConclusionActionRequired &&
-		check.ApplyID == 0 &&
-		check.BlockingReason == rollbackCompletedBlock.blockingReason &&
-		check.ErrorMessage == rollbackCompletedBlock.message
-}
-
-// TestE2ERollbackIgnoredByNonOwningInstance verifies that in a multi-instance
-// setup, an instance that doesn't own the apply's environment silently ignores
-// rollback commands instead of reacting or posting "Apply Not Found".
-func TestE2ERollbackIgnoredByNonOwningInstance(t *testing.T) {
-	dbName := "webhook_rb_multienv"
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"production"})
-	ctx := t.Context()
-
-	// Seed a completed apply for staging (owned by the other instance)
-	_, err := svc.Storage().Applies().Create(ctx, &storage.Apply{
-		ApplyIdentifier: "apply-aabbccdd0011",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           "completed",
-		Engine:          "spirit",
-	})
-	require.NoError(t, err)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	comments := make(chan string, 10)
-	reactions := make(chan string, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/1/comments", func(w http.ResponseWriter, r *http.Request) {
-		var body struct {
-			Body string `json:"body"`
-		}
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		comments <- body.Body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 99})
-	})
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/comments/42/reactions", func(w http.ResponseWriter, r *http.Request) {
-		var body struct {
-			Content string `json:"content"`
-		}
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		reactions <- body.Content
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	commands := []string{
-		"schemabot rollback apply-aabbccdd0011 -e staging",
-		"schemabot rollback-confirm apply-aabbccdd0011 -e staging",
-		"schemabot rollback-confirm -e staging",
-	}
-	for _, command := range commands {
-		req := buildWebhookRequest(t, webhookPayloadOpts{
-			comment: command,
-			isPR:    true,
-		}, nil)
-		rr := httptest.NewRecorder()
-		h.ServeHTTP(rr, req)
-		require.Equal(t, http.StatusOK, rr.Code)
-		assert.Contains(t, rr.Body.String(), "environment handled by another instance")
-	}
-
-	// The production instance should NOT post any comment (it silently ignores
-	// the staging commands). Wait long enough for any async handler to fire.
-	select {
-	case body := <-comments:
-		t.Fatalf("production instance should not post a comment for staging rollback command, got: %s", body)
-	case reaction := <-reactions:
-		t.Fatalf("production instance should not react to staging rollback command, got: %s", reaction)
-	case <-time.After(2 * time.Second):
-		// Expected: no comment or reaction posted.
-	}
-}
-
-// TestE2EPRCloseCleanup verifies PR-close cleanup against real storage. While
-// the PR has a running apply, closing it retains the database lock (so no other
-// PR can start a concurrent apply on the same database) and retains stored
-// check state (so a close-and-reopen cannot turn in-flight apply state into a
-// passing check). Once the apply reaches a terminal state, closing the PR
-// releases the lock and deletes the stored check state.
-func TestE2EPRCloseCleanup(t *testing.T) {
-	dbName := "webhook_pr_close"
-	svc := setupE2EService(t, dbName)
-
-	// Seed a lock and check record for this PR
-	err := svc.Storage().Locks().Acquire(t.Context(), &storage.Lock{
-		DatabaseName: dbName,
-		DatabaseType: "mysql",
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		Owner:        "octocat/hello-world#1",
-	})
-	require.NoError(t, err)
-
-	err = svc.Storage().Checks().Upsert(t.Context(), &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "abc123",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   42,
-		HasChanges:   true,
-		Status:       "completed",
-		Conclusion:   "action_required",
-	})
-	require.NoError(t, err)
-
-	applyID, err := svc.Storage().Applies().Create(t.Context(), &storage.Apply{
-		ApplyIdentifier: "apply-close-running",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           state.Apply.Running,
-		Engine:          "spirit",
-	})
-	require.NoError(t, err)
-
-	h := NewHandler(svc, nil, nil, slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError})))
-
-	// Close the PR while the apply is running. The handler is invoked directly
-	// (not through the async webhook goroutine) so the retention assertions
-	// below observe a finished cleanup pass.
-	h.handlePRClosed("octocat/hello-world", 1, 12345)
-
-	lock, err := svc.Storage().Locks().Get(t.Context(), dbName, "mysql")
-	require.NoError(t, err)
-	require.NotNil(t, lock, "lock must be retained while the apply is running")
-	assert.Equal(t, "octocat/hello-world#1", lock.Owner)
-
-	check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check, "stored check state must be retained while the apply is running")
-	assert.Equal(t, "action_required", check.Conclusion)
-
-	// Finish the apply, then close the PR again through the webhook path.
-	apply, err := svc.Storage().Applies().Get(t.Context(), applyID)
-	require.NoError(t, err)
-	require.NotNil(t, apply)
-	apply.State = state.Apply.Completed
-	require.NoError(t, svc.Storage().Applies().Update(t.Context(), apply))
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{action: "closed"}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "PR close cleanup started")
-
-	// Poll until lock is released (cleanup runs async)
-	require.Eventually(t, func() bool {
-		lock, err := svc.Storage().Locks().Get(t.Context(), dbName, "mysql")
-		return err == nil && lock == nil
-	}, 5*time.Second, 100*time.Millisecond, "lock should be released on PR close once the apply is terminal")
-
-	// Poll until check is deleted
-	require.Eventually(t, func() bool {
-		check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", dbName)
-		return err == nil && check == nil
-	}, 5*time.Second, 100*time.Millisecond, "check should be deleted on PR close once the apply is terminal")
-}
-
-// TestE2EStaleCheckCleanup verifies that checks for databases no longer in the PR
-// are marked as success on synchronize.
-func TestE2EStaleCheckCleanup(t *testing.T) {
-	dbName := "webhook_stale_check"
-	svc := setupE2EService(t, dbName)
-
-	// Seed a check for a database that WON'T be in the next auto-plan
-	err := svc.Storage().Checks().Upsert(t.Context(), &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "old-sha",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: "removed_database",
-		CheckRunID:   42,
-		HasChanges:   true,
-		Status:       "completed",
-		Conclusion:   "action_required",
-	})
-	require.NoError(t, err)
-
-	// Also seed a check for the database that WILL be in the auto-plan
-	err = svc.Storage().Checks().Upsert(t.Context(), &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "old-sha",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   43,
-		HasChanges:   true,
-		Status:       "completed",
-		Conclusion:   "action_required",
-	})
-	require.NoError(t, err)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	h := newE2EHandler(t, svc, client)
-
-	// Send synchronize event (simulates new commits pushed)
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{action: "synchronize"}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// The stale check (removed_database) should be updated to success
-	require.Eventually(t, func() bool {
-		check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", "removed_database")
-		if err != nil || check == nil {
-			return false
-		}
-		return check.Conclusion == "success"
-	}, 10*time.Second, 200*time.Millisecond, "stale check should be updated to success")
-
-	// The active check (dbName) should still exist (may be updated by auto-plan)
-	check, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check, "active check should still exist")
-}
-
-// TestE2EReconcileStaleInProgressCheck verifies that when a check is stuck at
-// "in_progress" from a crashed apply, the next plan or apply command reconciles
-// it to the apply's terminal state. This prevents branch protection from being
-// blocked indefinitely after a server crash mid-apply.
-func TestE2EReconcileStaleInProgressCheck(t *testing.T) {
-	dbName := "webhook_stale_inprogress"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Seed a completed apply (simulates an apply that finished but the goroutine died)
-	apply := &storage.Apply{
-		ApplyIdentifier: "apply-stale-test",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           state.Apply.Completed,
-		Engine:          "spirit",
-	}
-	applyID, err := svc.Storage().Applies().Create(ctx, apply)
-	require.NoError(t, err)
-	apply.ID = applyID
-
-	// Seed a check stuck at "in_progress" (the goroutine died before updating it)
-	err = svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha999",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   42,
-		ApplyID:      applyID,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	})
-	require.NoError(t, err)
-
-	// Seed an aggregate also stuck at in_progress
-	err = svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha999",
-		Environment:  aggregateSentinel,
-		DatabaseType: aggregateSentinel,
-		DatabaseName: aggregateSentinel,
-		CheckRunID:   100,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	})
-	require.NoError(t, err)
-
-	// Set up fake GitHub
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	result := setupFakeGitHubForPlan(t, mux, nil, "", dbName)
-	h := newE2EHandler(t, svc, client)
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-
-	require.NoError(t, h.reconcileStaleChecks(ctx, installClient, "octocat/hello-world", 1))
-
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, checkStatusCompleted, check.Status)
-	assert.Equal(t, checkConclusionSuccess, check.Conclusion)
-	assert.Equal(t, applyID, check.ApplyID)
-
-	select {
-	case checkRun := <-result.checkRuns:
-		assert.Equal(t, aggregateCheckName, checkRun.Name)
-		assert.Equal(t, "abc123", checkRun.HeadSHA)
-		assert.Equal(t, checkStatusCompleted, checkRun.Status)
-		assert.Equal(t, checkConclusionSuccess, checkRun.Conclusion)
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for aggregate check run")
-	}
-
-	aggregate, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	require.NotNil(t, aggregate)
-	assert.Equal(t, "abc123", aggregate.HeadSHA)
-	assert.Equal(t, checkStatusCompleted, aggregate.Status)
-	assert.Equal(t, checkConclusionSuccess, aggregate.Conclusion)
-}
-
-// TestE2EReconcileStaleInProgressCheckFailure verifies startup/webhook
-// reconciliation for a check that is still in_progress even though its apply
-// already failed. Reconciliation must publish the failure instead of leaving
-// branch protection pending forever.
-func TestE2EReconcileStaleInProgressCheckFailure(t *testing.T) {
-	dbName := "webhook_stale_inprogress_failed"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Seed the terminal apply. This models a driver that reached a failed apply
-	// state, then crashed before it updated GitHub check state.
-	apply := &storage.Apply{
-		ApplyIdentifier: "apply-stale-failed",
-		Database:        dbName,
-		DatabaseType:    "mysql",
-		Environment:     "staging",
-		Repository:      "octocat/hello-world",
-		PullRequest:     1,
-		State:           state.Apply.Failed,
-		Engine:          "spirit",
-	}
-	applyID, err := svc.Storage().Applies().Create(ctx, apply)
-	require.NoError(t, err)
-	apply.ID = applyID
-
-	// The stored per-database and aggregate check state still say in_progress.
-	// The per-database row points at the failed apply that reconciliation should
-	// use as the source of truth.
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha999",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		CheckRunID:   42,
-		ApplyID:      applyID,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha999",
-		Environment:  aggregateSentinel,
-		DatabaseType: aggregateSentinel,
-		DatabaseName: aggregateSentinel,
-		CheckRunID:   100,
-		HasChanges:   true,
-		Status:       checkStatusInProgress,
-		Conclusion:   "",
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// Fake GitHub provides current PR metadata so reconciliation can safely
-	// update the aggregate check on the current commit SHA.
-	result := setupFakeGitHubForPlan(t, mux, nil, "", dbName)
-	h := newE2EHandler(t, svc, client)
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-
-	require.NoError(t, h.reconcileStaleChecks(ctx, installClient, "octocat/hello-world", 1))
-
-	// Reconciliation should copy the failed apply result into the stored
-	// per-database check while preserving ownership by the original apply.
-	check, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, check)
-	assert.Equal(t, checkStatusCompleted, check.Status)
-	assert.Equal(t, checkConclusionFailure, check.Conclusion)
-	assert.Equal(t, applyID, check.ApplyID)
-
-	select {
-	case checkRun := <-result.checkRuns:
-		assert.Equal(t, aggregateCheckName, checkRun.Name)
-		assert.Equal(t, "abc123", checkRun.HeadSHA)
-		assert.Equal(t, checkStatusCompleted, checkRun.Status)
-		assert.Equal(t, checkConclusionFailure, checkRun.Conclusion)
-	case <-time.After(webhookIntegrationPollDeadline):
-		t.Fatal("timed out waiting for failing aggregate check run")
-	}
-}
-
-// TestE2EMultiAppAutoPlan simulates a monorepo with multiple apps, each with their
-// own schema directory and database. Verifies that a PR touching one app only creates
-// checks for that app's database, not for others.
-func TestE2EMultiAppAutoPlan(t *testing.T) {
-	// Create two databases simulating two apps in a monorepo
-	paymentsSvc := setupE2EService(t, "payments")
-	ordersSvc := setupE2EService(t, "orders")
-	_ = ordersSvc // orders is configured but not touched by this PR
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// Simulate a monorepo with two apps, each with their own schema dir:
-	//   payments-service/mysql/schema/schemabot.yaml  → database: payments
-	//   orders-service/mysql/schema/schemabot.yaml    → database: orders
-	// The PR only changes payments — orders should NOT be planned.
-	paymentsConfig := "database: payments\ntype: mysql\n"
-	ordersConfig := "database: orders\ntype: mysql\n"
-	transactionsSQL := "CREATE TABLE `transactions` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `amount_cents` bigint NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-	auditLogSQL := "CREATE TABLE `audit_log` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `action` varchar(50) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;"
-
-	result := &planFlowResult{
-		comments:  make(chan string, 10),
-		reactions: make(chan string, 10),
-		checkRuns: make(chan checkRunCapture, 10),
-	}
-
-	// PR info
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("abc123")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files in payments-service only (two namespaces: payments + payments_audit)
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("payments-service/mysql/schema/payments/transactions.sql"), Status: new("added")},
-			{Filename: new("payments-service/mysql/schema/payments_audit/audit_log.sql"), Status: new("added")},
-		})
-	})
-
-	// Git tree contains BOTH apps' schema files (full repo tree)
-	treeEntries := []*gh.TreeEntry{
-		// payments app (two namespaces)
-		{Path: new("payments-service/mysql/schema/schemabot.yaml"), Mode: new("100644"), Type: new("blob"), SHA: new("configsha_payments"), Size: new(len(paymentsConfig))},
-		{Path: new("payments-service/mysql/schema/payments/transactions.sql"), Mode: new("100644"), Type: new("blob"), SHA: new("blobsha_transactions"), Size: new(len(transactionsSQL))},
-		{Path: new("payments-service/mysql/schema/payments_audit/audit_log.sql"), Mode: new("100644"), Type: new("blob"), SHA: new("blobsha_audit"), Size: new(len(auditLogSQL))},
-		// orders app (not in changed files)
-		{Path: new("orders-service/mysql/schema/schemabot.yaml"), Mode: new("100644"), Type: new("blob"), SHA: new("configsha_orders"), Size: new(len(ordersConfig))},
-	}
-
-	blobContents := map[string]string{
-		"configsha_payments":   paymentsConfig,
-		"blobsha_transactions": transactionsSQL,
-		"blobsha_audit":        auditLogSQL,
-		"configsha_orders":     ordersConfig,
-	}
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/abc123", func(w http.ResponseWriter, r *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{SHA: new("abc123"), Entries: treeEntries, Truncated: new(false)})
-	})
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/blobs/", func(w http.ResponseWriter, r *http.Request) {
-		sha := r.URL.Path[len("/repos/octocat/hello-world/git/blobs/"):]
-		if _, ok := blobContents[sha]; !ok {
-			http.NotFound(w, r)
-			return
-		}
-		encoded := base64.StdEncoding.EncodeToString([]byte(blobContents[sha]))
-		w.WriteHeader(http.StatusOK)
-		_, _ = fmt.Fprintf(w, `{"sha":"%s","content":"%s","encoding":"base64","size":%d}`, sha, encoded, len(blobContents[sha]))
-	})
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/contents/", func(w http.ResponseWriter, r *http.Request) {
-		filePath := r.URL.Path[len("/repos/octocat/hello-world/contents/"):]
-		if filePath == "payments-service/mysql/schema/schemabot.yaml" {
-			_ = json.NewEncoder(w).Encode(gh.RepositoryContent{
-				Name:     new("schemabot.yaml"),
-				Path:     new("payments-service/mysql/schema/schemabot.yaml"),
-				Content:  new(base64.StdEncoding.EncodeToString([]byte(paymentsConfig))),
-				Encoding: new("base64"),
-			})
-			return
-		}
-		http.NotFound(w, r)
-	})
-
-	// Capture comments, reactions, check runs
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/1/comments", func(w http.ResponseWriter, r *http.Request) {
-		var body struct {
-			Body string `json:"body"`
-		}
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		result.comments <- body.Body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 99})
-	})
-	mux.HandleFunc("POST /repos/octocat/hello-world/issues/comments/42/reactions", func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		result.checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-	mux.HandleFunc("PATCH /repos/octocat/hello-world/check-runs/", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		result.checkRuns <- body
-		w.WriteHeader(http.StatusOK)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	// Wire up BOTH databases in the service
-	h := newE2EHandler(t, paymentsSvc, client)
-
-	// Send PR opened webhook (triggers auto-plan)
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{action: "opened"}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Should get a plan comment for payments only, showing both namespaces
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "payments", "plan comment should be for payments database")
-		assert.NotContains(t, body, "orders", "should NOT plan orders database")
-		assert.Contains(t, body, "CREATE TABLE")
-		assert.Contains(t, body, "transactions", "should include payments namespace table")
-		assert.Contains(t, body, "audit_log", "should include payments_audit namespace table")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for plan comment")
-	}
-
-	// Only the aggregate check run should be created.
-	hasAggregate := false
-	deadline := time.After(5 * time.Second)
-	for {
-		select {
-		case cr := <-result.checkRuns:
-			if cr.Name == aggregateCheckName {
-				hasAggregate = true
-				goto checksDone
-			}
-		case <-deadline:
-			goto checksDone
-		}
-	}
-checksDone:
-	assert.True(t, hasAggregate, "expected aggregate check run")
-}
-
 func e2eContainerName(base string) string {
 	out, err := exec.CommandContext(context.Background(), "git", "rev-parse", "--abbrev-ref", "HEAD").Output()
 	if err != nil {
@@ -3694,86 +716,6 @@ func e2eContainerName(base string) string {
 		return '-'
 	}, branch)
 	return base + "-" + sanitized
-}
-
-func TestE2EWebhookMetrics(t *testing.T) {
-	// Set up OTel ManualReader to capture metrics.
-	reader := sdkmetric.NewManualReader()
-	mp := sdkmetric.NewMeterProvider(sdkmetric.WithReader(reader))
-	prevMP := otel.GetMeterProvider()
-	otel.SetMeterProvider(mp)
-	t.Cleanup(func() {
-		otel.SetMeterProvider(prevMP)
-		require.NoError(t, mp.Shutdown(t.Context()))
-	})
-
-	dbName := "webhook_metrics_test"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// Send an issue_comment webhook (plan command).
-	commentReq := buildWebhookRequest(t, webhookPayloadOpts{
-		comment: "schemabot plan -e staging",
-		isPR:    true,
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, commentReq)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Send a pull_request opened webhook (auto-plan trigger).
-	prReq := buildPRWebhookRequest(t, prWebhookPayloadOpts{action: "opened"}, nil)
-	rr2 := httptest.NewRecorder()
-	h.ServeHTTP(rr2, prReq)
-	require.Equal(t, http.StatusOK, rr2.Code)
-
-	// Collect metrics and verify both webhook events were recorded.
-	var rm metricdata.ResourceMetrics
-	require.NoError(t, reader.Collect(t.Context(), &rm))
-
-	var found bool
-	observedEvents := make(map[string]bool)
-	for _, sm := range rm.ScopeMetrics {
-		for _, m := range sm.Metrics {
-			if m.Name == "schemabot.webhook.events_total" {
-				found = true
-				sum, ok := m.Data.(metricdata.Sum[int64])
-				require.True(t, ok)
-
-				for _, dp := range sum.DataPoints {
-					evType, _ := dp.Attributes.Value(attribute.Key("event_type"))
-					action, _ := dp.Attributes.Value(attribute.Key("action"))
-					status, _ := dp.Attributes.Value(attribute.Key("status"))
-					repo, _ := dp.Attributes.Value(attribute.Key("repository"))
-					key := evType.AsString() + "/" + action.AsString()
-					observedEvents[key] = true
-					t.Logf("webhook metric: event_type=%s action=%s repo=%s status=%s",
-						evType.AsString(), action.AsString(), repo.AsString(), status.AsString())
-				}
-			}
-		}
-	}
-	require.True(t, found, "schemabot.webhook.events_total metric not found")
-	assert.True(t, observedEvents["issue_comment/created"], "expected issue_comment/created metric")
-	assert.True(t, observedEvents["pull_request/opened"], "expected pull_request/opened metric")
 }
 
 // setupE2EServiceWithAllowedEnvs creates a service with AllowedEnvironments set.
@@ -3802,772 +744,4 @@ func setupE2EServiceWithAllowedEnvs(t *testing.T, allowedEnvs []string) *api.Ser
 	t.Cleanup(func() { _ = svc.Close() })
 
 	return svc
-}
-
-// TestE2EAutoPlanFailsWhenConfiguredEnvironmentsAreNotAllowed verifies that
-// SchemaBot fails closed when schema files changed, but the server-configured
-// database environments do not overlap this service's allowed environments.
-// This is a configuration mismatch, not "no work".
-func TestE2EAutoPlanFailsWhenConfiguredEnvironmentsAreNotAllowed(t *testing.T) {
-	dbName := "webhook_no_owned_envs"
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"sandbox"})
-	configureE2EServiceEnvironments(t, svc, dbName, "staging", "production")
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// The database is configured for staging and production, but this test service
-	// processes only sandbox. SchemaBot cannot safely plan this schema change
-	// because none of the configured environments are allowed.
-	schemabotConfig := fmt.Sprintf("database: %s\ntype: mysql\n", dbName)
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, dbName)
-	h := newE2EHandler(t, svc, client)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "auto-plan started")
-
-	// The webhook still accepts the PR event asynchronously, but auto-plan posts
-	// a failing aggregate because this service cannot process any environment
-	// configured for the database.
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, "SchemaBot (sandbox)", cr.Name)
-		assert.Equal(t, "abc123", cr.HeadSHA)
-		assert.Equal(t, checkStatusCompleted, cr.Status)
-		assert.Equal(t, checkConclusionFailure, cr.Conclusion)
-		require.NotNil(t, cr.Output)
-		assert.Equal(t, noAllowedConfiguredEnvironmentsBlock.message, cr.Output.Summary)
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for failing aggregate for allowed environment")
-	}
-
-	require.Eventually(t, func() bool {
-		aggregate, err := svc.Storage().Checks().Get(t.Context(), "octocat/hello-world", 1, "sandbox", aggregateSentinel, aggregateSentinel)
-		return err == nil && aggregate != nil &&
-			aggregate.Conclusion == checkConclusionFailure &&
-			aggregate.BlockingReason == noAllowedConfiguredEnvironmentsBlock.blockingReason &&
-			aggregate.ErrorMessage == noAllowedConfiguredEnvironmentsBlock.message
-	}, 5*time.Second, 100*time.Millisecond, "failing aggregate should be stored for allowed environment")
-
-	// There should be no plan comment because no environment reached planning.
-	select {
-	case body := <-result.comments:
-		t.Fatalf("expected no plan comment when no configured environments are allowed, got: %s", body)
-	case <-time.After(500 * time.Millisecond):
-	}
-}
-
-// TestE2EPassingAggregateOnNonSchemaPR verifies that when a PR doesn't touch schema
-// files and allowed_environments is configured, passing aggregate checks are posted
-// so branch protection isn't blocked.
-func TestE2EPassingAggregateOnNonSchemaPR(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging", "production"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR info
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("abc123")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files — no schema files
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-		})
-	})
-
-	// Git tree — no schemabot.yaml
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/abc123", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA:       new("abc123"),
-			Entries:   []*gh.TreeEntry{},
-			Truncated: new(false),
-		})
-	})
-
-	// Capture check runs
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "no schema files in PR")
-
-	// Wait for both passing aggregates (staging + production)
-	seen := map[string]bool{}
-	for i := range 2 {
-		select {
-		case cr := <-checkRuns:
-			seen[cr.Name] = true
-			assert.Equal(t, "completed", cr.Status)
-			assert.Equal(t, "success", cr.Conclusion)
-		case <-time.After(10 * time.Second):
-			t.Fatalf("timed out waiting for passing aggregate check run %d/2, seen: %v", i+1, seen)
-		}
-	}
-	assert.True(t, seen["SchemaBot (staging)"], "expected SchemaBot (staging) check")
-	assert.True(t, seen["SchemaBot (production)"], "expected SchemaBot (production) check")
-}
-
-// TestE2ECheckRunRerequestReplansCurrentPR verifies that rerunning a SchemaBot
-// Check Run from GitHub reuses auto-plan discovery and republishes aggregate
-// check state for the current PR head.
-func TestE2ECheckRunRerequestReplansCurrentPR(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("abc123")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/abc123", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA:       new("abc123"),
-			Entries:   []*gh.TreeEntry{},
-			Truncated: new(false),
-		})
-	})
-
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-	req := buildCheckRunWebhookRequest(t, checkRunWebhookPayloadOpts{
-		checkName: "SchemaBot (staging)",
-		headSHA:   "abc123",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "no schema files in PR")
-
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, "SchemaBot (staging)", cr.Name)
-		assert.Equal(t, "abc123", cr.HeadSHA)
-		assert.Equal(t, checkStatusCompleted, cr.Status)
-		assert.Equal(t, checkConclusionSuccess, cr.Conclusion)
-	case <-time.After(webhookIntegrationCheckRunDeadline):
-		t.Fatal("timed out waiting for rerun aggregate check")
-	}
-}
-
-// TestE2ECheckRunRerequestIgnoresStaleHeadSHA verifies that rerunning an old
-// Check Run cannot publish check state for a commit that is no longer the PR
-// head.
-func TestE2ECheckRunRerequestIgnoresStaleHeadSHA(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	var fileListCalls atomic.Int64
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("newsha222")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		fileListCalls.Add(1)
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-		})
-	})
-
-	checkRuns := make(chan checkRunCapture, 1)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-	req := buildCheckRunWebhookRequest(t, checkRunWebhookPayloadOpts{
-		checkName: "SchemaBot (staging)",
-		headSHA:   "oldsha111",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "stale head SHA")
-	assert.Equal(t, int64(0), fileListCalls.Load(), "stale check reruns should stop before config discovery")
-
-	select {
-	case cr := <-checkRuns:
-		t.Fatalf("stale check rerun should not publish a check run, got: %+v", cr)
-	case <-time.After(250 * time.Millisecond):
-	}
-}
-
-// TestE2EPassingAggregateSynchronizeUpdatesNewSHA verifies that when a non-schema PR
-// receives a synchronize event (force push / new commit), the passing aggregate check
-// is recreated on the new HEAD SHA — not left stale on the old commit.
-func TestE2EPassingAggregateSynchronizeUpdatesNewSHA(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	var currentHead atomic.Value
-	currentHead.Store("sha1aaa")
-
-	// PR info. This endpoint represents the current head SHA for the PR, so
-	// update it before sending each webhook event.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		headSHA := currentHead.Load().(string)
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: &headSHA},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files — no schema files
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-		})
-	})
-
-	// Git tree — no schemabot.yaml
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/sha1aaa", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA: new("sha1aaa"), Entries: []*gh.TreeEntry{}, Truncated: new(false),
-		})
-	})
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/sha2bbb", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA: new("sha2bbb"), Entries: []*gh.TreeEntry{}, Truncated: new(false),
-		})
-	})
-
-	// Capture check runs with HEAD SHA
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	// Step 1: PR opened with sha1
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "sha1aaa",
-	}, nil)
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, "sha1aaa", cr.HeadSHA, "first aggregate should be on the opened SHA")
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for check run on opened event")
-	}
-
-	// Step 2: synchronize with sha2 (force push)
-	currentHead.Store("sha2bbb")
-	req = buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "synchronize",
-		headSHA: "sha2bbb",
-	}, nil)
-	rr = httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, "sha2bbb", cr.HeadSHA, "aggregate must be recreated on the new SHA after synchronize")
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out — aggregate was not recreated on new SHA after synchronize")
-	}
-}
-
-// TestE2EAggregateUpdateSkipsStaleHeadSHA verifies that aggregate updates are
-// gated by the current PR commit SHA. A stale driver must not publish a check
-// run for an older commit after the PR branch has moved.
-func TestE2EAggregateUpdateSkipsStaleHeadSHA(t *testing.T) {
-	dbName := "webhook_stale_sha_guard"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	// Seed per-database check state from an older commit. The aggregate update
-	// below will try to use this old SHA.
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "oldsha111",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		HasChanges:   true,
-		Status:       checkStatusCompleted,
-		Conclusion:   checkConclusionActionRequired,
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// GitHub reports that the PR is now on a newer commit.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{
-				Ref: new("feature-branch"),
-				SHA: new("newsha222"),
-			},
-			Base: &gh.PullRequestBranch{
-				Ref: new("main"),
-				SHA: new("def456"),
-			},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	checkRuns := make(chan checkRunCapture, 1)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-	h.updateAggregateCheck(ctx, installClient, "octocat/hello-world", 1, "oldsha111")
-
-	// The old-SHA aggregate update should be skipped entirely: no GitHub check
-	// run and no stored aggregate row.
-	select {
-	case cr := <-checkRuns:
-		t.Fatalf("stale aggregate update should not publish a check run, got: %+v", cr)
-	case <-time.After(250 * time.Millisecond):
-	}
-
-	aggregate, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	assert.Nil(t, aggregate)
-}
-
-// TestE2EDisabledRepoChecksSkipAggregatePublishing verifies that a server-side
-// repository safety hatch suppresses GitHub Check Runs while preserving stored
-// per-database check state for SchemaBot's own safety decisions.
-func TestE2EDisabledRepoChecksSkipAggregatePublishing(t *testing.T) {
-	dbName := "webhook_disabled_repo_checks"
-	svc := setupE2EService(t, dbName)
-	ctx := t.Context()
-
-	enableChecks := false
-	svc.Config().Repos = map[string]api.RepoConfig{
-		"octocat/hello-world": {EnableChecks: &enableChecks},
-	}
-
-	require.NoError(t, svc.Storage().Checks().Upsert(ctx, &storage.Check{
-		Repository:   "octocat/hello-world",
-		PullRequest:  1,
-		HeadSHA:      "abc123",
-		Environment:  "staging",
-		DatabaseType: "mysql",
-		DatabaseName: dbName,
-		HasChanges:   true,
-		Status:       checkStatusCompleted,
-		Conclusion:   checkConclusionActionRequired,
-	}))
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	var githubCalls atomic.Int64
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		githubCalls.Add(1)
-		w.WriteHeader(http.StatusInternalServerError)
-	})
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, _ *http.Request) {
-		githubCalls.Add(1)
-		w.WriteHeader(http.StatusInternalServerError)
-	})
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-	h := newE2EHandler(t, svc, client)
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-
-	h.updateAggregateCheck(ctx, installClient, "octocat/hello-world", 1, "abc123")
-	h.postPassingAggregates(ctx, installClient, "octocat/hello-world", 1, "abc123")
-	h.postFailingAggregates(ctx, installClient, "octocat/hello-world", 1, "abc123", map[string]string{
-		"staging": "Plan failed",
-	})
-
-	assert.Equal(t, int64(0), githubCalls.Load(), "disabled check publishing should not call GitHub")
-
-	storedCheck, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", "mysql", dbName)
-	require.NoError(t, err)
-	require.NotNil(t, storedCheck, "per-database check state should remain available")
-	assert.Equal(t, checkConclusionActionRequired, storedCheck.Conclusion)
-
-	aggregate, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, aggregateSentinel, aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	assert.Nil(t, aggregate, "disabled check publishing should not store aggregate check state")
-}
-
-// TestE2EPassingAggregateRequiresGitHubHeadVerification verifies that passing
-// aggregate paths still verify the current PR commit before publishing a check.
-func TestE2EPassingAggregateRequiresGitHubHeadVerification(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-	ctx := t.Context()
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// The helper cannot verify the PR head because GitHub is unavailable.
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-	})
-
-	checkRuns := make(chan checkRunCapture, 1)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	h := newE2EHandler(t, svc, client)
-	installClient := ghclient.NewInstallationClient(client, h.logger)
-	h.postPassingAggregates(ctx, installClient, "octocat/hello-world", 1, "abc123")
-
-	// Without head verification, SchemaBot must not publish or store a passing
-	// aggregate check that could incorrectly unblock branch protection.
-	select {
-	case cr := <-checkRuns:
-		t.Fatalf("passing aggregate should not publish without current head verification, got: %+v", cr)
-	case <-time.After(250 * time.Millisecond):
-	}
-
-	aggregate, err := svc.Storage().Checks().Get(ctx, "octocat/hello-world", 1, "staging", aggregateSentinel, aggregateSentinel)
-	require.NoError(t, err)
-	assert.Nil(t, aggregate)
-}
-
-// TestE2EPassingAggregateOnSQLWithoutSchemabotYAML verifies that when a PR touches
-// .sql files but the directory has no schemabot.yaml (not onboarded), passing
-// aggregate checks are still posted.
-func TestE2EPassingAggregateOnSQLWithoutSchemabotYAML(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR info
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("abc123")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files — .sql file but in a directory without schemabot.yaml
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("legacy-service/schema/users.sql"), Status: new("modified")},
-		})
-	})
-
-	// Git tree — has the .sql file but no schemabot.yaml anywhere
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/abc123", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA: new("abc123"),
-			Entries: []*gh.TreeEntry{
-				{
-					Path: new("legacy-service/schema/users.sql"),
-					Mode: new("100644"),
-					Type: new("blob"),
-					SHA:  new("blobsha001"),
-					Size: new(100),
-				},
-			},
-			Truncated: new(false),
-		})
-	})
-
-	// Capture check runs
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "no schema files in PR")
-
-	// Should post passing aggregate even though .sql files changed
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, "SchemaBot (staging)", cr.Name)
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for passing aggregate check run")
-	}
-}
-
-// TestE2EPassingAggregateWithoutAllowedEnvs verifies that when allowed_environments
-// is not configured (single instance mode), a global "SchemaBot" passing aggregate
-// is posted for non-schema PRs.
-func TestE2EPassingAggregateWithoutAllowedEnvs(t *testing.T) {
-	dbName := "webhook_no_aggregate"
-	svc := setupE2EService(t, dbName)
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// PR info
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.PullRequest{
-			Head: &gh.PullRequestBranch{Ref: new("feature-branch"), SHA: new("abc123")},
-			Base: &gh.PullRequestBranch{Ref: new("main"), SHA: new("def456")},
-			User: &gh.User{Login: new("testuser")},
-		})
-	})
-
-	// PR changed files — no schema files
-	mux.HandleFunc("GET /repos/octocat/hello-world/pulls/1/files", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode([]*gh.CommitFile{
-			{Filename: new("README.md"), Status: new("modified")},
-		})
-	})
-
-	// Git tree — empty
-	mux.HandleFunc("GET /repos/octocat/hello-world/git/trees/abc123", func(w http.ResponseWriter, _ *http.Request) {
-		_ = json.NewEncoder(w).Encode(gh.Tree{
-			SHA:       new("abc123"),
-			Entries:   []*gh.TreeEntry{},
-			Truncated: new(false),
-		})
-	})
-
-	// Capture check runs
-	checkRuns := make(chan checkRunCapture, 10)
-	mux.HandleFunc("POST /repos/octocat/hello-world/check-runs", func(w http.ResponseWriter, r *http.Request) {
-		var body checkRunCapture
-		_ = json.NewDecoder(r.Body).Decode(&body)
-		checkRuns <- body
-		w.WriteHeader(http.StatusCreated)
-		_ = json.NewEncoder(w).Encode(map[string]any{"id": 1})
-	})
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code)
-	assert.Contains(t, rr.Body.String(), "no schema files in PR")
-
-	// Single-instance mode posts a global "SchemaBot" passing aggregate
-	select {
-	case cr := <-checkRuns:
-		assert.Equal(t, "SchemaBot", cr.Name)
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "success", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for passing aggregate check run")
-	}
-}
-
-// TestE2EFailingAggregateOnPlanError verifies that when a plan fails for all
-// environments (e.g., database not configured), a failing aggregate check is
-// posted so branch protection shows a clear failure instead of waiting forever.
-func TestE2EFailingAggregateOnPlanError(t *testing.T) {
-	svc := setupE2EServiceWithAllowedEnvs(t, []string{"staging"})
-
-	mux := http.NewServeMux()
-	server := httptest.NewServer(mux)
-	t.Cleanup(server.Close)
-
-	client := gh.NewClient(nil)
-	client.BaseURL, _ = url.Parse(server.URL + "/")
-
-	// schemabot.yaml references a database not configured on the server
-	schemabotConfig := "database: unconfigured-db\ntype: mysql\n"
-	schemaFiles := map[string]string{
-		"users.sql": "CREATE TABLE `users` (\n  `id` bigint unsigned NOT NULL AUTO_INCREMENT,\n  `name` varchar(255) NOT NULL,\n  PRIMARY KEY (`id`)\n) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;",
-	}
-
-	result := setupFakeGitHubForPlan(t, mux, schemaFiles, schemabotConfig, "unconfigured-db")
-
-	logger := slog.New(slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelError}))
-	installClient := ghclient.NewInstallationClient(client, logger)
-	factory := &fakeClientFactory{client: installClient}
-
-	h := NewHandler(svc, factory, nil, logger)
-
-	req := buildPRWebhookRequest(t, prWebhookPayloadOpts{
-		action:  "opened",
-		headSHA: "abc123",
-		headRef: "feature-branch",
-	}, nil)
-
-	rr := httptest.NewRecorder()
-	h.ServeHTTP(rr, req)
-	require.Equal(t, http.StatusOK, rr.Code)
-
-	// Should get a comment with the error
-	select {
-	case body := <-result.comments:
-		assert.Contains(t, body, "Error")
-		assert.Contains(t, body, "unconfigured-db")
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for error comment")
-	}
-
-	// Should also get a failing aggregate check run
-	select {
-	case cr := <-result.checkRuns:
-		assert.Equal(t, "SchemaBot (staging)", cr.Name)
-		assert.Equal(t, "completed", cr.Status)
-		assert.Equal(t, "failure", cr.Conclusion)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for failing aggregate check run")
-	}
 }
